requirements.txt
kafka-python
redis
scikit-learn
numpy
fastapi
uvicorn
anthropic
shap
pandas
